Renamed InterfaceWithDatabase class to be more explicit
